Designate action:Action and core:Event disjoint
In an OWL-only sense, the ontology portion of this patch could have been
accomplished by adding one triple to `action:Action`.

For the sake of symmetry and explicitness, `core:Event` also picked up
the `owl:disjointWith` statement.  To satisfy OWL syntactic
requirements, a stub reference to `action:Action` was added, as was done
for `types:Dictionary` for Issue 541 (discussed in PR 542).

To enforce disjointedness with SHACL, a new independent shape is added,
`action:Action-disjointWith-Event-shape`.  The shape is spelled using an
anonymous node shape to avoid a multiple-inheritance issue:

* SHACL syntactically requires the subject of a triple with `sh:not` as
  predicate be a blank node.
* Separately, SHACL also only permits a shape to have up to one triple
  with `sh:not` as predicate.

Hence, if the `action.ttl` file were included twice in some graph,
SHACL-SHACL validation would fail, because the identically-written blank
nodes that are `sh:not` objects would not resolve to the same node.
Using an anonymous wrapping `sh:NodeShape` avoids this issue.
(Multiple inheritance would now cause some redundant processing to occur,
but SHACL-SHACL validation would pass.)

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ucoProject/UCO/issues/563
* https://github.com/ucoProject/UCO/pull/542
* https://www.w3.org/TR/shacl/#shacl-shacl

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>